serving/requirements.txt
torch
transformers
flask
google-cloud-storage
gunicorn
